+++ src/main/resources/allProperties.properties
@@ -1,1 +1,1 @@
-propFiles=global.properties,objectRepo1.properties,objectRepo2.properties
+propFiles=global.properties,objectRepo1.properties
