+++ src/main/resources/objectRepo1.properties
@@ -13,6 +13,8 @@ Customer=Xpath://bdi[contains(text()\,'Customer')]
 Search=Xpath://*[@id='menuItem_businesspartner.search-unifiedmenu-txt' or @id='tab1-businessPartnerSearch-search--search-execute-button-content' ] 
 Person_Origination=Xpath://div[@id='menuItem_businesspartner.create-unifiedmenu-txt']
 Organization_Origination=Xpath://div[@id='menuItem_org.create-unifiedmenu-txt']
+
+#Customer Origination
 TIN_SSN_EIN=Xpath://input[@id='tab1-businessPartnerSearch-search--searchByBasicData-tin-textInput-inner']
 Spinner=Xpath://div[@id='spinner']
 Customer_no_column=Xpath://td[@class="sapUiTableTd"][2]//span
@@ -70,7 +72,6 @@ actual_name=//div[contains(@id\,'businessPartnerOrigination-ProcessSummary--Proc
 actual_Birth_Date=//div[contains(@id\,'businessPartnerOrigination-ProcessSummary--ProcessSummary-form')]//span[contains(text()\,'Birth Date')]/../following-sibling::div[1]//span
 actual_Email_address=//div[contains(@id\,'businessPartnerOrigination-ProcessSummary--ProcessSummary-form')]//span[contains(text()\,'Email address')]/../following-sibling::div[1]//span
 actual_address=//div[contains(@id\,'businessPartnerOrigination-ProcessSummary--ProcessSummary-form')]//span[contains(text()\,'Address')]/../following-sibling::div[1]//span
-
 Create_Organization=Xpath://bdi[contains(text()\,'Create Organization')]
 Entity_name=Xpath://input[contains(@id\,'businessPartnerOrigination-captureOrganization--masterData-businessName')]
 Organization_type=Xpath://div[contains(@id\,'businessPartnerOrigination-captureOrganization--masterData-orgType')]
