--- src/main/resources/objectRepo2.properties
@@ -1,1 +1,0 @@
-vast.button.text2 = submit2
