+++ src/test/java/features/OrganizationOrigination.feature
@@ -12,7 +12,7 @@ Background:
     And user clicks on button "Login"
     And user is on "Dashboard_page"    
  	
-@Smoke @Regression	 @TC003
+@Smoke @Regression	 @TC007
 Scenario Outline: Validate the user is able to create new record for organization in VAST system when no customer is retreived from the system
 	Given user is on "Dashboard_page" 
 	And user clicks on button "Customer"
@@ -67,7 +67,7 @@ Scenario Outline: Validate the user is able to create new record for organizatio
 
     Examples: 
 	|	TIN (SSN/EIN)	|	Entity name	|	Organization type	|	Registered state |	Street address	|	City	|	State		|	Zip code	|	Phone type	|	Phone number	|	Email address	|	ID type						|	Number	|	Issuing authority	|	ID state	|	Issuance date	|	Expiration date	|	Officer code	|	Branch	     | Relation type | Customer  | Address                       |
-	|	114454411		|	VAST Entity	|	Corporation			|	Oklahoma 		 |	1234 Main St	|	Tulsa	|	Oklahoma	|	12345		|	Cell		|	987-654-3210	|	abc@abc.com		|	Articles of Incorporation	|	DMV1234	|	DMV					|	Texas		|	Jun 07, 2016	|	Jun 05, 2022	|	RBS             |	Tulsa - Yale | Trustee       | 444444444 | 1234 Main St, Tulsa, OK 12345 |	
+	|	114464611		|	VAST Entity	|	Corporation			|	Oklahoma 		 |	1234 Main St	|	Tulsa	|	Oklahoma	|	12345		|	Cell		|	987-654-3210	|	abc@abc.com		|	Articles of Incorporation	|	DMV1234	|	DMV					|	Texas		|	Jun 07, 2016	|	Jun 05, 2022	|	RBS             |	Tulsa - Yale | Trustee       | 444444444 | 1234 Main St, Tulsa, OK 12345 |
